@@ -23,8 +23,8 @@
     "@enact/spotlight": "^2.2.9",
     "@enact/ui": "^2.2.9",
     "@enact/webos": "^2.2.9",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2"
+    "react": "^16.7.0",
+    "react-dom": "^16.7.0"
   },
   "devDependencies": {
     "@enact/dev-utils": "^0.4.0",
