@@ -28,7 +28,7 @@
   "dependencies": {
     "@enact/core": "^2.2.9",
     "prop-types": "^15.6.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0"
+    "react": "^16.7.0",
+    "react-dom": "^16.7.0"
   }
 }
