@@ -11,6 +11,53 @@ import computed from './computed';
 import styles from './styles';
 
 /**
+ * @callback RenderFunction
+ * @memberof core/kind
+ * @param {Object<string, any>} props
+ * @param {Object<string, any>} context
+ * @returns React.Element|null
+ */
+
+/**
+ * @callback ComputedPropFunction
+ * @memberof core/kind
+ * @param {Object<string, any>} props
+ * @param {Object<string, any>} context
+ * @returns any
+ */
+
+/**
+ * @callback HandlerFunction
+ * @memberof core/kind
+ * @param {any} event
+ * @param {Object<string, any>} props
+ * @param {Object<string, any>} context
+ */
+
+/**
+ * Configuration for CSS class name mapping
+ *
+ * @typedef {Object} StylesBlock
+ * @memberof core/kind
+ * @property {Object.<string, string>} css
+ * @property {String} [className]
+ * @property {Boolean|String|String[]} [publicClassNames]
+ */
+
+/**
+ * @typedef {Object} KindConfig
+ * @memberof core/kind
+ * @property {String} name
+ * @property {Object.<string, Function>} [propTypes]
+ * @property {Object.<string, any>} [defaultProps]
+ * @property {Object.<string, Function>} [contextTypes]
+ * @property {StylesBlock} [styles]
+ * @property {Object.<string, HandlerFunction>} [handlers]
+ * @property {Object.<string, ComputedPropFunction>} [computed]
+ * @property {RenderFunction} render
+ */
+
+/**
  * Creates a new component with some helpful declarative syntactic sugar.
  *
  * Example:
@@ -55,7 +102,7 @@ import styles from './styles';
  * ```
  *
  * @function
- * @param  {Object}    config    Component configuration
+ * @param  {KindConfig}    config    Component configuration
  *
  * @returns {Function}           Component
  * @memberof core/kind
