@@ -1,5 +1,5 @@
 import React from 'react';
-import {mount, shallow} from 'enzyme';
+import {mount} from 'enzyme';
 import {EditableIntegerPicker, EditableIntegerPickerBase} from '../EditableIntegerPicker';
 import Spotlight from '@enact/spotlight';
 
@@ -161,11 +161,11 @@ describe('EditableIntegerPicker', () => {
 	});
 
 	test('should be disabled when limited to a single value', () => {
-		const picker = shallow(
+		const picker = mount(
 			<EditableIntegerPickerBase min={0} max={0} value={0} />
 		);
 
-		const actual = picker.find('SpottablePicker').last().prop('disabled');
+		const actual = picker.find('Picker').last().prop('disabled');
 		expect(actual).toBe(true);
 	});
 
