@@ -9,7 +9,7 @@
  * />
  *
  * @module moonstone/GridListImageItem
- * @exports GridListIamgeItem
+ * @exports GridListImageItem
  * @exports GridListImageItemBase
  * @exports GridListImageItemDecorator
  */
@@ -52,15 +52,6 @@ const GridListImageItemBase = kind({
 
 	propTypes: /** @lends moonstone/GridListImageItem.GridListImageItemBase.prototype */ {
 		/**
-		 * The voice control intent.
-		 *
-		 * @type {String}
-		 * @default 'Select'
-		 * @public
-		 */
-		'data-webos-voice-intent': PropTypes.string,
-
-		/**
 		 * Customizes the component by mapping the supplied collection of CSS class names to the
 		 * corresponding internal Elements and states of this component.
 		 *
@@ -76,6 +67,16 @@ const GridListImageItemBase = kind({
 		 * @public
 		 */
 		css: PropTypes.object,
+
+		/**
+		 * The voice control intent.
+		 *
+		 * @type {String}
+		 * @default 'Select'
+		 * @memberof moonstone/GridListImageItem.GridListImageItemBase.prototype
+		 * @public
+		 */
+		'data-webos-voice-intent': PropTypes.string,
 
 		/**
 		 * Placeholder image used while [source]{@link ui/GridListImageItem.GridListImageItem#source}
@@ -181,7 +182,7 @@ const GridListImageItemDecorator = compose(
  *
  * @class GridListImageItem
  * @memberof moonstone/GridListImageItem
- * @extends moonstone/GridListIamgeItem.GridListImageItemBase
+ * @extends moonstone/GridListImageItem.GridListImageItemBase
  * @mixes moonstone/GridListImageItem.GridListImageItemDecorator
  * @see moonstone/GridListImageItem.GridListImageItemBase
  * @ui
