@@ -1,5 +1,5 @@
 import React from 'react';
-import {mount, shallow} from 'enzyme';
+import {mount} from 'enzyme';
 import {RangePicker, RangePickerBase} from '../RangePicker';
 
 const tap = (node) => {
@@ -70,11 +70,11 @@ describe('RangePicker Specs', () => {
 	});
 
 	test('should be disabled when limited to a single value', () => {
-		const picker = shallow(
+		const picker = mount(
 			<RangePickerBase min={0} max={0} value={0} />
 		);
 
-		const actual = picker.find('SpottablePicker').last().prop('disabled');
+		const actual = picker.find('Picker').last().prop('disabled');
 		expect(actual).toBe(true);
 	});
 });
