@@ -5,7 +5,7 @@ import css from '../Spinner.less';
 
 describe('Spinner Specs', () => {
 	test(
-		'should have not have client node when Spinner has no children',
+		'should not have client node when Spinner has no children',
 		() => {
 			const spinner = mount(
 				<Spinner />
@@ -49,7 +49,7 @@ describe('Spinner Specs', () => {
 		() => {
 			const spinner = mount(
 				<Spinner transparent>
-                    Loading...
+					Loading...
 				</Spinner>
 			);
 
