@@ -68,7 +68,8 @@ import {
 	getTargetByContainer,
 	getTargetByDirectionFromElement,
 	getTargetByDirectionFromPosition,
-	getTargetBySelector
+	getTargetBySelector,
+	isFocusable
 } from './target';
 
 import {
@@ -511,6 +512,8 @@ const Spotlight = (function () {
 		 * Initializes Spotlight. This is generally handled by
 		 * {@link spotlight/SpotlightRootDecorator.SpotlightRootDecorator}.
 		 *
+		 * @param {Object} containerDefaults Default configuration for new spotlight containers
+		 * @returns {undefined}
 		 * @public
 		 */
 		initialize: function (containerDefaults) {
@@ -573,8 +576,9 @@ const Spotlight = (function () {
 		 * Sets the config for spotlight or the specified containerID
 		 *
 		 * @function
-		 * @param {String|Object} param1 Configuration object or container ID
-		 * @param {Object|undefined} param2 Configuration object if container ID supplied in param1
+		 * @param {String|Object} containerIdOrConfig  Configuration object or container ID
+		 * @param {Object}        [config]             Configuration object if container ID supplied
+		 *                                             in `containerIdOrConfig`
 		 * @returns {undefined}
 		 * @public
 		 */
@@ -587,8 +591,9 @@ const Spotlight = (function () {
 		 * object. If no container ID is supplied, a new container ID will be generated.
 		 *
 		 * @function
-		 * @param {String|Object} param1 Configuration object or container ID
-		 * @param {Object|undefined} param2 Configuration object if container ID supplied in param1
+		 * @param {String|Object} containerIdOrConfig  Configuration object or container ID
+		 * @param {Object}        [config]             Configuration object if container ID supplied
+		 *                                             in `containerIdOrConfig`
 		 * @returns {String} The container ID of the container
 		 * @public
 		 */
@@ -759,7 +764,8 @@ const Spotlight = (function () {
 		 * Sets or clears the default container that will receive focus.
 		 *
 		 * @function
-		 * @param {String|undefined} containerId The container ID or a falsy value to clear default container
+		 * @param {String} [containerId] The container ID or a falsy value to clear default
+		 *                               container
 		 * @returns {undefined}
 		 * @public
 		 */
@@ -846,13 +852,13 @@ const Spotlight = (function () {
 				return false;
 			}
 
-			return matchSelector('.' + spottableClass, elem);
+			return isFocusable(elem);
 		},
 
 		/**
 		 * Returns the currently spotted control.
 		 *
-		 * @returns {Object} The control that currently has focus, if available
+		 * @returns {Node} The control that currently has focus, if available
 		 * @public
 		 */
 		getCurrent: function () {
@@ -862,8 +868,8 @@ const Spotlight = (function () {
 		/**
 		 * Returns a list of spottable elements wrapped by the supplied container.
 		 *
-		 * @param {String} [containerId] The id of the container used to determine the list of spottable elements
-		 * @returns {NodeList} The spottable elements that are wrapped by the supplied container
+		 * @param {String} containerId The id of the container used to determine the list of spottable elements
+		 * @returns {Node[]} The spottable elements that are wrapped by the supplied container
 		 * @public
 		 */
 		getSpottableDescendants: function (containerId) {
